Meer

+++ bblaser-arm/src/domain/color.cpp
@@ -2,22 +2,24 @@
 // Created by Gijs Sijpesteijn on 14/01/2021.
 //
 
+
+#include <ostream>
 #include "color.h"
 
-color::color(int red, int green, int blue) {
+color::color(u_int8_t red, u_int8_t green, u_int8_t blue) {
     this->red = red;
     this->green = green;
     this->blue = blue;
 }
 
-int color::getRed() const {
+u_int8_t color::getRed() const {
     return this->red;
 }
 
-int color::getGreen() const {
+u_int8_t color::getGreen() const {
     return this->green;
 }
 
-int color::getBlue() const {
+u_int8_t color::getBlue() const {
     return this->blue;
 }
